+++ module/Task/src/Domain/Task/Exception/TaskAlreadyCompleted.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Task\Domain\Task\Exception;
+
+use DomainException;
+
+final class TaskAlreadyCompleted extends DomainException
+{
+    public function __construct()
+    {
+        parent::__construct('Task is already completed.');
+    }
+}